refactor: wrong names and style
