+++ frontend/src/config.ts
@@ -1,1 +1,2 @@
 export const BACKEND_URL = import.meta.env.VITE_BACKEND_URL;
+export const RED = "#9F2D20"
