+++ frontend/src/main.tsx
@@ -12,8 +12,6 @@ import LoginPage from "./pages/LoginPage.tsx";
 import AboutPage from "./pages/AboutPage.tsx";
 import CommandsPage from "./pages/CommandsPage.tsx";
 
-export const RED = "#9F2D20"
-
 const router = createBrowserRouter([
 
     {
